chore(rpc): cleanup
