+++ crates/node2/Cargo.toml
@@ -56,6 +56,7 @@ metrics-exporter-prometheus = { workspace = true }
 sysinfo = { workspace = true, features = ["system", "network"] }
 
 # Other
+libp2p-identity = { workspace = true }
 xxhash-rust = { workspace = true }
 axum = { workspace = true }
 envy = { workspace = true }
@@ -76,6 +77,7 @@ tracing-subscriber = { workspace = true, features = [
     "json",
 ] }
 
+libp2p-identity = { workspace = true, features = ["rand"] }
 alloy = { workspace = true, features = [
     "provider-anvil-node",
     "node-bindings",
