+++ deny.toml
@@ -18,11 +18,3 @@ allow = [
 
 [licenses.private]
 ignore = true
-
-# TODO We should be able to remove `ignore-sources` once we add `publish = false` to all the crates sourced from here.
-ignore-sources = ["https://github.com/WalletConnect/utils-rs.git"]
-
-[[licenses.clarify]]
-name = "ring"
-expression = "ISC"
-license-files = [{ path = "LICENSE", hash = 0xbd0eed23 }]
